apply /31 and /32 fix from interface_add() to interface_modify()
This code looks messy but it is a exact copy of the code that was
previously added to interface_add() to allow /31 and /32 subnets.
It should have also been added to interface_modify() so you could
modify /31 nd /32 subnets.
Without this you cannot change the name of a /31 or /32 subnet.